requirements.txt
aniso8601==7.0.0
django==2.2.7
django-filter==2.2.0
django-graphql-jwt==0.2.2
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
